safari-600.7-branch/LayoutTests/platform/win/js/dom/global-constructors-attributes-shared-worker-expected.txt

Test to ensure that global constructors have the right attributes in the shared worker context

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: script-tests/global-constructors-attributes.js
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Blob').value is Blob
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Blob').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Blob').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Blob').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Blob').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'EventSource').value is EventSource
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'EventSource').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'EventSource').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'EventSource').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'EventSource').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessageChannel').value is MessageChannel
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessageChannel').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessageChannel').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessageChannel').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessageChannel').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessageEvent').value is MessageEvent
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessageEvent').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessageEvent').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessageEvent').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessageEvent').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'SharedWorkerGlobalScope').value is SharedWorkerGlobalScope
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'SharedWorkerGlobalScope').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'SharedWorkerGlobalScope').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'SharedWorkerGlobalScope').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'SharedWorkerGlobalScope').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'URL').value is URL
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'URL').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'URL').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'URL').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'URL').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WebSocket').value is WebSocket
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WebSocket').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WebSocket').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WebSocket').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WebSocket').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerGlobalScope').value is WorkerGlobalScope
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerGlobalScope').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerGlobalScope').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerGlobalScope').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerGlobalScope').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerLocation').value is WorkerLocation
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerLocation').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerLocation').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerLocation').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerLocation').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'XMLHttpRequest').value is XMLHttpRequest
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'XMLHttpRequest').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'XMLHttpRequest').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'XMLHttpRequest').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'XMLHttpRequest').configurable is true
PASS successfullyParsed is true

TEST COMPLETE
